Allow an array step and Callable object learning rate.
Callable objects have an additional `self` parameter, so we could have
a learning rate with two inputs that still works as
```
learning_rate(step)
```
For example, this is how Keras `LearningRateSchedule`s work.  Modified
the learning rate code to allow for this.

Also allow a `jax.Array` for the step count.  Training steps in Keras
are stored in a keras variable backed by a 0-D `jax.Array`.  We can
pass this directly to the learning rate function.

PiperOrigin-RevId: 752440907