Remove hard restriction that padded embedding dims match in explicit stack_tables.
This is not a hard-requirement from the implementation itself.  The main issue is that
it may lead to excess padding.  Note that `auto_stack_tables` already handles this by
automatically grouping tables into groups with the same padded embedding dim.

Reduced the exception to a logged warning.  If a user wants to explicitly stack
specific tables, they should be able to.  This is useful in tests when generating
randomized tables to stack.

PiperOrigin-RevId: 752758206