[JAX SC] [23/n] [Minibatching] Add gRPC-based AllReduce implementation for SparseCore.
This change introduces a new gRPC service and client interface to perform all-reduce operations across multiple tasks. The `AllReduceGrpcService` allows hosts to contribute data for a given synchronization key, and the `GrpcAllReduceInterface` orchestrates the all-to-all exchange and reduction of data. The current implementation supports boolean (logical OR) and uint64 (bitwise OR) reductions.

PiperOrigin-RevId: 798336901