[JAX SC] [21/n] [Minibatching] Integrate `MinibatchingSyncService` into input preprocessing.
This change refactors the input preprocessing pipeline to use the new `MinibatchingSyncService`. The previous logic using `absl::Barrier` for on-host synchronization is replaced by the new service.

This also adds the `AllReduceInterface*` to the options struct, allowing the preprocessing function to perform cross-host collectives. As part of this, the return type of `PreprocessSparseDenseMatmulInput` is changed to `absl::StatusOr` to handle potential errors from collective operations, and callers have been updated accordingly.

PiperOrigin-RevId: 799314358