[JAX SC] Refactor: Remove MinibatchingSyncService from input preprocessing
This change refactors the input preprocessing logic to remove the `MinibatchingSyncService` class.

The synchronization of `minibatching_required` and `minibatching_split` values across threads is now handled by directly calling the `BlockingAllReduce` method on the `AllReduceInterface`. This simplifies the code by removing a layer of abstraction.

To support this, `Serialize` and `Deserialize` helper functions were added for `bool` and `MinibatchingSplit` types.

PiperOrigin-RevId: 805125280