[JAX SC] Refactor `all_reduce_test` to use `MinibatchingNode` test utils.
This change refactors `AllReduceTest` and `MultipleLocalValuesAllReduceTest` to use the `SetUpMinibatchingNodes` helper function. This simplifies the test setup by removing the manual creation of gRPC services, servers, and interfaces, and instead relying on the `MinibatchingNode` abstraction.

The `SetUpMinibatchingNodes` utility is updated to support configuring the number of threads per task.

A new test case for `BlockingAllReduce` with boolean values is also added to `MultipleLocalValuesAllReduceTest`.

PiperOrigin-RevId: 805540589