Explicitly set py/abi wheel tags and use Google org self-hosted runner.
The Google runner supports multithreading and larger disk space for the job.
But it doesn't have `clang` pre-installed, so our hack for forcing a
platform-specific wheel to set the py/abi wheel tags no longer works.
We work around this by explicitly setting the tags using the python
`wheel` package.

PiperOrigin-RevId: 828499821