Add shared ownership to `tstring`'s VIEW type.
This change introduces `tstring::owner<T>`, a reference-counted wrapper for an object `T`. `tstring` can now be assigned a "shared view" using `assign_as_shared_view`, where the underlying data is owned by a `tstring::owner`. The `tstring` will increment the owner's reference count and decrement it when the `tstring` is deallocated or reassigned, ensuring the owner object remains alive as long as any `tstring` holds a shared view to its data. This is implemented by adding a `TStringOwnerCApi` pointer to the `TF_TString_View` struct and updating the relevant `TF_TString` functions to handle the reference counting.

PiperOrigin-RevId: 832101409