[JAX SC] Cast `embedding_table.type` to `ir.RankedTensorType` before calling `get_dim_size`.
`embedding_table.type` is an instance of `ir.Type`, but `get_dim_size` is only available on subclasses like `ir.RankedTensorType`. While `maybe_downcast()` attempts to cast to a subtype, it seems static analysis tools like pylint are unable to infer the correct subtype and still raise an 'attribute-error' as shown in the screenshot.

Explicitly casting with `ir.RankedTensorType(embedding_table.type)` ensures that we get an instance of `RankedTensorType`, which is known to have the `get_dim_size` method, thus resolving the error and removing the need for pylint suppression comments.

PiperOrigin-RevId: 833528462